Move find_executable to utils (#1799)
We don't want to import the whole patroni.ctl into the patroni
